+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `clock` module for monotonic wall-time
 - Add `crypto::cbc` module for AES-256-CBC
 - Add `usb::ctap` module for CTAP HID
 - Improve safety documentation